+++ Tests/Unit/Service/UserServiceTest.php
@@ -1,0 +1,60 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AOE\Crawler\Tests\Unit\Service;
+
+/*
+ * (c) 2020 AOE GmbH <dev@aoe.com>
+ *
+ * This file is part of the TYPO3 Crawler Extension.
+ *
+ * It is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU General Public License, either version 2
+ * of the License, or any later version.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ *
+ * The TYPO3 project - inspiring people to share!
+ */
+
+use AOE\Crawler\Service\UserService;
+use Nimut\TestingFramework\TestCase\UnitTestCase;
+
+class UserServiceTest extends UnitTestCase
+{
+    /**
+     * @test
+     *
+     * @dataProvider hasGroupAccessDataProvider
+     */
+    public function hasGroupAccess(string $groupList, string $accessList, bool $expected): void
+    {
+        self::assertEquals(
+            $expected,
+            UserService::hasGroupAccess($groupList, $accessList)
+        );
+    }
+
+    public function hasGroupAccessDataProvider(): array
+    {
+        return [
+            'Do not have access' => [
+                'groupList' => '1,2,3',
+                'accessList' => '4,5,6',
+                'expected' => false,
+            ],
+            'Do have access' => [
+                'groupList' => '1,2,3,4',
+                'accessList' => '4,5,6',
+                'expected' => true,
+            ],
+            'Access List empty' => [
+                'groupList' => '1,2,3',
+                'accessList' => '',
+                'expected' => true,
+            ],
+        ];
+    }
+}